# openapi.yaml
openapi: 3.0.0
info:
  title: Second Brain Server API
  description: |
    Comprehensive personal intelligence platform API for knowledge management, creativity, learning, and personal growth.

    ## Features
    - **Stateless Authentication**: JWT-based authentication with no server-side sessions
    - **OAuth Integration**: Google OAuth support with CSRF protection
    - **Database Management**: Create and manage custom databases with flexible schemas
    - **Record Management**: CRUD operations for database records with rich property types
    - **User Management**: Complete user lifecycle management with role-based access control
    - **Rate Limiting**: Built-in rate limiting for security and performance
    - **File Operations**: Import/export data in multiple formats (JSON, CSV, Excel)
    - **Real-time Collaboration**: Share databases with granular permissions

    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    - Login: 5 attempts per 15 minutes
    - Registration: 3 attempts per hour
    - Password Reset: 3 attempts per hour
    - Token Refresh: 10 attempts per 15 minutes
    - OAuth: 20 attempts per 15 minutes
    - General API: 100 requests per 15 minutes

    ## Database Property Types
    - **TEXT**: Single-line text
    - **NUMBER**: Numeric values
...
    - **DATE**: Date and time values
...
  version: 1.0.0
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          type: object
          description: Response data (varies by endpoint)
        error:
          type: object
          properties:
            message:
              type: string
              example: Error description
            statusCode:
              type: integer
              example: 400
            status:
              type: string
              example: error
...
    WorkspaceListResponse:
      type: object
      properties:
        workspaces:
          type: array
          items:
            $ref: '#/components/schemas/Workspace'
          description: List of workspaces
        total:
          type: integer
          description: Total number of workspaces
          example: 25
        totalPages:
          type: integer
          description: Total number of pages
          example: 3
        currentPage:
          type: integer
          description: Current page number
          example: 1
        hasNextPage:
          type: boolean
          description: Whether there are more pages
          example: true
        hasPrevPage:
          type: boolean
          description: Whether there are previous pages
          example: false
...
paths:
...
  /workspaces:
    get:
      tags: [Workspace Management]
      summary: Get user's workspaces
      description: |
        Retrieve workspaces that the user owns or is a member of, with filtering and pagination support.

        **Features**:
        - Filter by role, public status, tags
        - Search by name and description
        - Sort by various criteria
        - Pagination support
      operationId: getUserWorkspaces
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          description: Search query for workspace name/description
          schema:
            type: string
            maxLength: 100
          example: "team"
        - name: role
          in: query
          description: Filter by user's role in workspace
          schema:
            type: string
            enum: [owner, admin, editor, viewer]
          example: "owner"
        - name: isPublic
          in: query
          description: Filter by public status
          schema:
            type: boolean
          example: false
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
          example: "development,team"
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [name, createdAt, updatedAt, lastActivityAt, memberCount, databaseCount]
            default: updatedAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Workspaces retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkspaceListResponse'
              examples:
                success:
                  summary: Successful response
                  value:
                    success: true
                    message: Workspaces retrieved successfully
                    data:
                      workspaces:
                        - id: 507f1f77bcf86cd799439030
                          name: Team Alpha
                          description: Main workspace for Team Alpha
                          icon: 🚀
                          ownerId: 507f1f77bcf86cd799439013
                          isPublic: false
                          memberCount: 8
                          databaseCount: 15
                          lastActivityAt: 2023-01-01T10:30:00Z
                      total: 12
                      totalPages: 1
                      currentPage: 1
                      hasNextPage: false
                      hasPrevPage: false
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    post:
      tags: [Workspace Management]
      summary: Create a new workspace
      description: |
        Create a new workspace. The user becomes the owner and first member.

        **Features**:
        - Automatic owner membership
        - Customizable settings
        - Tag support
        - Color themes
      operationId: createWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
            examples:
              basic_workspace:
                summary: Basic workspace
                value:
                  name: Team Alpha
                  description: Main workspace for Team Alpha projects
                  icon: 🚀
              full_workspace:
                summary: Full workspace with all options
                value:
                  name: Development Team
                  description: Workspace for development projects
                  icon: 💻
                  color: "#3B82F6"
                  isPublic: false
                  allowMemberInvites: true
                  defaultDatabasePermission: read
                  tags: ["development", "team"]
      responses:
        '201':
          description: Workspace created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workspace'
              examples:
                success:
                  summary: Workspace created
                  value:
                    success: true
                    message: Workspace created successfully
                    data:
                      id: 507f1f77bcf86cd799439030
                      name: Team Alpha
                      description: Main workspace for Team Alpha projects
                      icon: 🚀
                      ownerId: 507f1f77bcf86cd799439013
                      isPublic: false
                      memberCount: 1
                      databaseCount: 0
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '409':
          description: Workspace name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /workspaces/public:
    get:
      tags: [Workspace Management]
      summary: Get public workspaces
      description: |
        Retrieve publicly accessible workspaces. No authentication required.

        **Features**:
        - Browse public workspaces
        - Search and filter capabilities
        - Pagination support
      operationId: getPublicWorkspaces
      parameters:
        - name: search
          in: query
          description: Search query
          schema:
            type: string
            maxLength: 100
        - name: tags
          in: query
          description: Filter by tags
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Public workspaces retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkspaceListResponse'
  /workspaces/search:
    get:
      tags: [Workspace Management]
      summary: Search workspaces
      description: |
        Search workspaces with advanced filtering options.

        **Features**:
        - Full-text search
        - Include public workspaces option
        - Tag filtering
        - Relevance-based sorting
      operationId: searchWorkspaces
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: includePublic
          in: query
          description: Include public workspaces in results
          schema:
            type: boolean
            default: true
        - name: tags
          in: query
          description: Filter by tags
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum results
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkspaceListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /workspaces/stats:
    get:
      tags: [Workspace Management]
      summary: Get workspace statistics
      description: |
        Get comprehensive statistics about user's workspaces.

        **Features**:
        - Workspace counts by type
        - Database and member statistics
        - Recent activity summary
      operationId: getWorkspaceStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkspaceStats'
              examples:
                success:
                  summary: Workspace statistics
                  value:
                    success: true
                    message: Workspace statistics retrieved successfully
                    data:
                      totalWorkspaces: 12
                      ownedWorkspaces: 5
                      memberWorkspaces: 7
                      publicWorkspaces: 25
                      totalDatabases: 45
                      totalMembers: 23
                      recentActivity: []
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /workspaces/{id}:
    get:
      tags: [Workspace Management]
      summary: Get workspace by ID
      description: |
        Retrieve detailed information about a specific workspace.

        **Features**:
        - Complete workspace details
        - Member information
        - Database count
        - Permission checking
      operationId: getWorkspaceById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Workspace ID
          schema:
            type: string
          example: 507f1f77bcf86cd799439030
      responses:
        '200':
          description: Workspace retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workspace'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    put:
      tags: [Workspace Management]
      summary: Update workspace
      description: |
        Update workspace information and settings.

        **Features**:
        - Update basic information
        - Modify privacy settings
        - Change appearance settings
        - Update member permissions
      operationId: updateWorkspace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Workspace ID
          schema:
            type: string
          example: 507f1f77bcf86cd799439030
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkspaceRequest'
      responses:
        '200':
          description: Workspace updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workspace'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags: [Workspace Management]
      summary: Delete workspace
      description: |
        Delete a workspace permanently. Only workspace owners can delete workspaces.

        **Security Features**:
        - Owner-only access
        - Database dependency check
        - Permanent deletion warning
      operationId: deleteWorkspace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Workspace ID
          schema:
            type: string
          example: 507f1f77bcf86cd799439030
      responses:
        '200':
          description: Workspace deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /workspaces/{id}/members:
    get:
      tags: [Workspace Members]
      summary: Get workspace members
      description: |
        Retrieve list of workspace members with user details and roles.

        **Features**:
        - Member list with user information
        - Role-based filtering
        - Search by user details
        - Pagination support
      operationId: getWorkspaceMembers
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Workspace ID
          schema:
            type: string
          example: 507f1f77bcf86cd799439030
        - name: role
          in: query
          description: Filter by member role
          schema:
            type: string
            enum: [owner, admin, editor, viewer]
        - name: search
          in: query
          description: Search members by name or email
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Members retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkspaceMembersListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    post:
      tags: [Workspace Members]
      summary: Add member to workspace
      description: |
        Invite a user to join the workspace with specified role.

        **Features**:
        - Email or user ID invitation
        - Role assignment
        - Permission validation
        - Duplicate member check
      operationId: addWorkspaceMember
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Workspace ID
          schema:
            type: string
          example: 507f1f77bcf86cd799439030
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteMemberRequest'
      responses:
        '200':
          description: Member added successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workspace'
        '400':
          description: Validation error or user already member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /workspaces/{id}/members/{memberId}:
    delete:
      tags: [Workspace Members]
      summary: Remove member from workspace
      description: |
        Remove a member from the workspace. Admins can remove any member except owner.

        **Features**:
        - Permission-based removal
        - Owner protection
        - Self-removal allowed
        - Member notification
      operationId: removeWorkspaceMember
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Workspace ID
          schema:
            type: string
          example: 507f1f77bcf86cd799439030
        - name: memberId
          in: path
          required: true
          description: Member user ID
          schema:
            type: string
          example: 507f1f77bcf86cd799439014
      responses:
        '200':
          description: Member removed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workspace'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Insufficient permissions or cannot remove owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /workspaces/{id}/members/{memberId}/role:
    put:
      tags: [Workspace Members]
      summary: Update member role
      description: |
        Update a workspace member's role. Only admins and owners can change roles.

        **Features**:
        - Role hierarchy enforcement
        - Owner role protection
        - Permission validation
        - Role change notification
      operationId: updateMemberRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Workspace ID
          schema:
            type: string
          example: 507f1f77bcf86cd799439030
        - name: memberId
          in: path
          required: true
          description: Member user ID
          schema:
            type: string
          example: 507f1f77bcf86cd799439014
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRoleRequest'
      responses:
        '200':
          description: Member role updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workspace'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /workspaces/{id}/leave:
    post:
      tags: [Workspace Members]
      summary: Leave workspace
      description: |
        Leave a workspace (remove yourself as a member). Owners cannot leave their own workspace.

        **Features**:
        - Self-removal from workspace
        - Owner protection
        - Automatic cleanup
        - Member notification
      operationId: leaveWorkspace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Workspace ID
          schema:
            type: string
          example: 507f1f77bcf86cd799439030
      responses:
        '200':
          description: Left workspace successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Successfully left workspace
                  value:
                    success: true
                    message: Left workspace successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Owners cannot leave their own workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...